refactor: move signer to core
